fix: change getCommonFilms params from Integer to Long

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -54,7 +54,7 @@ public class FilmController {
     }
 
     @GetMapping("/common")
-    public List<Film> getCommonFilms(@RequestParam Integer userId, @RequestParam Integer friendId) {
+    public List<Film> getCommonFilms(@RequestParam Long userId, @RequestParam Long friendId) {
         return filmService.getCommonFilms(userId, friendId);
     }
 }
